requirements.txt
absl-py==0.2.0
astor==0.6.2
bleach==1.5.0
gast==0.2.0
grpcio==1.11.0
html5lib==0.9999999
Keras==2.1.5
Markdown==2.6.11
numpy==1.14.2
opencv-python==3.4.0.12
pandas==0.22.0
Pillow==5.1.0
protobuf==3.5.2.post1
python-dateutil==2.7.2
pytz==2018.4
PyYAML==3.12
scikit-learn==0.19.1
scipy==1.0.1
six==1.11.0
sklearn==0.0
tensorboard==1.7.0
tensorflow==1.7.0
termcolor==1.1.0
tflearn==0.3.2
Werkzeug==0.14.1
python-decouple